Bugfix as index should be int, not float

+++ raw_trainer/distribute_data.py
@@ -18,10 +18,10 @@ def distribute_plain_data():
 
     random.shuffle(file_paths)
     
-    train_index = num_files * proportion_train
-    validation_index = train_index + (num_files * proportion_validation)
-    test_index = validation_index + (num_files * proportion_test)
-    
+    train_index = int(num_files * proportion_train)
+    validation_index = int(train_index + (num_files * proportion_validation))
+    test_index = int(validation_index + (num_files * proportion_test))
+
     return file_paths[0:train_index], file_paths[train_index:validation_index], file_paths[validation_index:test_index]
 
 
@@ -40,8 +40,8 @@ def distribute_encrypted_data():
 
     random.shuffle(file_paths)
 
-    train_index = num_files * proportion_train
-    validation_index = train_index + (num_files * proportion_validation)
-    test_index = validation_index + (num_files * proportion_test)
+    train_index = int(num_files * proportion_train)
+    validation_index = int(train_index + (num_files * proportion_validation))
+    test_index = int(validation_index + (num_files * proportion_test))
 
     return file_paths[0:train_index], file_paths[train_index:validation_index], file_paths[validation_index:test_index]
